+++ gitcommonsync/tests/test_ansible_module.py
@@ -1,45 +1,27 @@
 import os
+import shutil
+import subprocess
 import unittest
-import warnings
-
-from gitcommonsync._ansible_runner import run_ansible, ResultCallbackHandler
+import sys
+from copy import copy
 
 _SCRIPT_LOCATION = os.path.dirname(os.path.realpath(__file__))
-_ANSIBLE_TEST_ROLE_NAME = "ansible"
-
-
-class CustomResultCallbackHandler(ResultCallbackHandler):
-    """
-    Callback handler that creates sub-tests for each Ansible task.
-    """
-    def __init__(self, test_case: unittest.TestCase):
-        super().__init__()
-        self.test_case = test_case
-
-    def v2_runner_on_ok(self, result, **kwargs):
-        super().v2_runner_on_ok(result, **kwargs)
-        self._on_complete(result)
-
-    def v2_runner_on_failed(self, result, ignore_errors=False):
-        super().v2_runner_on_failed(result, ignore_errors=ignore_errors)
-        self._on_complete(result)
-
-    def _on_complete(self, result):
-        with self.test_case.subTest(task=result.task_name):
-            self.test_case.assertFalse(result.is_failed(), result._result)
+_PROJECT_ROOT = os.path.join(_SCRIPT_LOCATION, "../..")
+_ANSIBLE_TEST_PLAYBOOK_LOCATION = os.path.join(_SCRIPT_LOCATION, "ansible/site.yml")
+_ANSIBLE_PLAYBOOK_BINARY = shutil.which("ansible-playbook")
 
 
 class TestAnsibleModule(unittest.TestCase):
     """
     Tests runner of the Ansible tests.
     """
-    def setUp(self):
-        warnings.simplefilter("ignore", ResourceWarning)
-        warnings.simplefilter("ignore", DeprecationWarning)
-
     def test_ansible_execution(self):
-        run_ansible(roles=[os.path.join(_SCRIPT_LOCATION, _ANSIBLE_TEST_ROLE_NAME)],
-                    results_callback_handler=CustomResultCallbackHandler(self))
+        environment = copy(os.environ)
+        environment["PROJECT_ROOT"] = _PROJECT_ROOT
+
+        subprocess.check_call([_ANSIBLE_PLAYBOOK_BINARY, "-i", "localhost,", "-c", "local", "-e",
+                               f"ansible_python_interpreter={sys.executable}", _ANSIBLE_TEST_PLAYBOOK_LOCATION],
+                              env=environment)
 
 
 if __name__ == "__main__":