+++ test-runner.sh
@@ -1,13 +1,13 @@
 #!/usr/bin/env bash
 set -eu -o pipefail
 
-export PROJECT_ROOT=$PWD
+export PYTHON_COVERAGE=1
 PYTHONPATH=. coverage run -m unittest discover -v -s gitcommonsync/tests
 
 coverage run setup.py install
 
 # Awful bit of munging to map coverage to the module in the project package
-sed -i -e "s#[^\"]*ansible_module.py#${PROJECT_ROOT}\/gitcommonsync\/ansible_module.py#g" .coverage*
+sed -i "" -e "s#[^\"]*ansible_module.py#${PWD}\/gitcommonsync\/ansible_module.py#g" .coverage*
 
 coverage combine -a
 coverage report