+++ CHANGELOG.md
@@ -1,26 +1,35 @@
 # Change Log
+## [Unreleased]
+### Changed
+- Calling Ansible via CLI opposed to (unsupported) Python API.
+
 
 ## 2.0.0 - 2017-08-08
 ### Changed
 - `commmitter_name` -> `author_name`; `commmitter_email` -> `author_email`.
 - Author name/email address is used when checking out subrepos.
 
+
 ## 1.2.0 - 2017-07-20
 ### Added
 - Support for working with newly created repositories, which have no commits.  
+
 
 ## 1.1.2 - 2017-07-20
 ### Changed
 - Minor improvements to address problems encountered when using the Ansible module in the wild.
 
+
 ## 1.1.1 - 2017-07-19
 ### Changed
 - Fixes PyPi installation problem.
+
 
 ## 1.1.0 - 2017-07-18
 ### Changed
 - Fixes Ansible import issue.
 - Changes visibility of repository module.
+
 
 ## 1.0.0 - 2017-07-18
 ### Added